requirements.txt
crewai
langchain
groq
streamlit
vercel
flask
flask-cors
flask-sqlalchemy
python-dotenv